+++ src/protocols/RoutingProtocolManagementServiceUserInterface.java
@@ -3,6 +3,6 @@ import java.util.SortedSet;
 
 public interface RoutingProtocolManagementServiceUserInterface {
 	void distanceTableIndication(short nodeId, SortedMap<Short, SortedSet<Double>> distanceTable);
-	void linkCostIndication(short nodeIdA, short nodeIdB, double cost);
+	void linkCostIndication(short nodeIdA, short nodeIdB, int cost);
 
 }
